fix local-aaa plugin loading
The local-aaa plugin will always attempt to load and generate an error
if there is no db configuration.   The log will return ```FTL failed to
connect to db error="error parsing uri: scheme must be \"mongodb\" or
\"mongodb+srv\""``` when this condition exists.

This commit will now check if the configuration file has a db uri and,
if it doesnt, will not load the plugin.