add support for user configuration in repositories
This commit adds two new configuration options for repositories for
setting the name and email address to add to the commit.  When set, the
commit message will now include those values so the committer can be
identified.

If the values are not set in the configuration file, emptry strings will
be used in the commit message.

To configure, simply add name and email to any repository object
configuration.

```
[repository test]
url = git@github.com:itential/assets.git
name = User Name
email = username@domain.com
```