fix repository push path
This commit fixes an issue when using the `--path` command line option
with the repository `push` command.  Prior to this change, the `--path`
command line option was ignored and therefore a user cannot push an
asset to a folder within the repo.

This change will now honor the `--path` command line option and place
the asset in the specified folder.

This commit also removes one of the options from the `pull` command.
There was an option to specify `--name` that should not have been there.
The option was ignored and effectively resulted in a noop.  It was
removed.