fixes issue trying to use --replace when importing automations (#4)
There is a bug in the automation import command that effectvely prevents
the use of the `--replace` command line option.  This commit addresses
that problem.  The `--replace` is now honored when importing
automations.

Additionally there where to command line options doing the same thing:
`--force` and `--replace`.  The application will standardize on
`--replace` going forward and this commit removes the `--force` option