fixes issues with jsonforms using both copy and import
This commit address two bugs.  The first issue was with jsonforms not
properly conforming to the copier interface.  It was missing the CopyTo
and CopyFrom functions.  This commit addresses that problem.

The second issue is with the JsonForm struct missing the version field.
Without that field, the jsonform would not properly import.  This issue
is fixed in this release.