fixes transformations to work with copy (#15)
This commit addresses an issue with the transformations runner not
adhering to the Copier interface.   This is now corrected and
transformations can be used with copy.

This commit also fixed a minor issue with deleting transformations.  The
expected status code was wrong in the service leading to a empty error.