fix import projects when using --expand (#37)
This commit fixes an issue when trying to import a project that was
exproted using `--expand`.  When the code was refactored to a common
`importProject` function, the path to the individual components was
lost.  This change fixes that issue such that import and pull now work
properly with expanded projects.