moves the local-aaa mongo url config param
This commit moves the location of the mongo url configuration parameter.
It was a global value and is now part of the profile.  This value must
be configured in order for the `local-aaa` plugin to work.

To set the value, add the following configuration value to any profile:

`mongo_url = mongodb://....`

Once added, the `local-aaa` plugin will be available.