display a more verbose error when importing a model
When importing a model, the application will check to see if the
workflow, pre and post transformations exist for every action defined in
the model.  It will return an error to the user if any one of the
dependencies does not exist.  The returned messages in some cases is
fairly generic.

The application will now return a more verbose error message to
indidcate which action the dependency failed to find prior to importing
the model.

fixes #67