Add feature flag support for dataset commands
One of the new features being considered for the application is support
for datasets.  Datasets provide bulk import and export of types of
assets.  This feature is under heavy consideration but is not ready for
production use; therefore, it is being feature flagged.

By default, dataset commands are disabled.  To enable dataset commands
add the following to your configuration:

```
[features]
datasets_enabled = [true, false*]
```

The default value for `datasets_enabled` is `false` which means the
command will not show up when running the application.  When set to
`true` the commands will show up.

Please use with caution as there are no guarantees this will be
available in future releases.