`export workflow --output json` returns "nul" as output (#84)
When running the `export workflow ...` command using the `--output json`
command line option, the application will return "null" to stdout.
This only happens when using the `--output json` flag (or if the default
output is set to `json`).  This fix will now return the actual workflow
object to stdout.

fixes #81