changes how import/export of prebuilts functions
When importing and exporting prebuilts, the application would create a
new key in the prebuilt called `filename`.   This change effectively
gets rid of that need and now just uses the `data` key.  When the
prebuilt as exported with `--expand`, the `data` key will now be
prefaces with `@` to denote the data is in an external file which is
consistent with the rest of the application.