The repository connections required ssh-agent
There is an issue in previous release where the application would ignore
any private key file path and simply rely on the ssh-agent to
authentiate to a private key repository.  This has now been fixed and
the application will now use a private key file specified in the
configuration or passed via the command line.

If there is no private key file specified, the application will fallback
to using the ssh-agent.  If the ssh-agent is not running, the
application will generate an error.

fixes #69