Fix issue with export profile api route
The is an error calling the `GET /profiles/{name}/export` API route.  The
server returns `Cannot read properties of undefined (reading 'length')`
error.

This issue is server side so this commit attempts to work around the
issue by calling the `GET /profiles/{name}` route instead for the export
function.

This change should be undone once the server side route is fixed.

fixes #92