The `import role ...` command fails with `--replace` on initial import (#98)
The `import role ...` command will fail with an error that states the
role does not exist when attempting to import a role for the first time
and using `--replace`.   This fix will now check the return message from
the API call and if the role doe not exist, it will not attempt to
replace it.

fixes #94