Add support for `--replace` to `import integration ...` command (#111)
The `import integration ...` command should be able to safely support
the `--replace` command line option.  This change will now replace an
existing integration of the same name if it exists.  Additionally, the
`import integration ...` command will now check if the model exists on
the server.  If it doesn't, the command will return an error.

fixes #107