Git name and email not properly reflect in commit message
When making a commit using the export command to a Git repository, the
configure name and email from the configuration file isn't properly
populating the git message.  That issue is addressed and the git commit
message is now properly updated.

This change also adds full unit testing for the runners repository
implementation.