Disable the connection timeout
The Platform client has a configurable parameter to set the timeout for
the connection.  There are times when it is advantageous to be able to
disable the connection timeout for long running jobs.  The connection
timeout can now be set to 0 which will effectively disable the
connection timeout and the job will run to completion.

This will also change the default value for the connection timeout.  The
old value was 5 seconds.  The new value is 0 seconds (disabled)