Fixes bug that causes an authorization token malformed error
The application returns `FTL error="The specified authorization token is
malformed or does not correspond with an active session"` error when
using oauth.   This is due to the token not be refreshed after the first
call to the application.  This was due to a conditional check on
authentication used for basic auth that should't have been applied to
oauth clients.

fixes #120