Expanding a project doesn't write the project file propertly
When exporting a project using the `--expand` comamnd line option, all
of the assets are propertly expanded to the path.  The main project file
though is not properly updated.  This results in a situation where by
tyring to import the expanded folder will seem like none of the
documents have changed.  This is now fixed in this patch.