Display warning on missing profile instead of exit (#131)
When running the application with a profile that is not configured, the
application will just error with no further message.   This prevents
someone from running `ipctl --help`.  The application will now display a
warning message if the profile is not found and configure the client
with all defaults.  This will allow the `ipctl --help` command to run
successfully.

fixes #129