Could not find profile with mixed case (#133)
When configuring the profile name with a mixed case for the
`default_profile` setting, the application would report the profile
could not be found.   Profile names are not case-sensitive and the
`default_profile` setting is.  This is now correct so the
`default_profile` setting is now case insensitive as well.

fixes #128